Build and release Brave and Chrome/Chromium bases
This commit supports building both Brave and Chrome/Chromium browser
bases. This is accomplished through a multi-stage Dockerfile which
builds the appropriate browser image based on the value of `--target`,
which can be either `brave` or `chrome`.

CI has been updated to:
- Run on every push of the `main` branch
- Run for both Brave and Chrome/Chromium separately
- Read the value of `brave-version.txt` and `chrome-version.txt` to
determine which version of each browser to build
- Create a draft release for each browser/version combination on GitHub
if one does not already exist

The build scripts have also been updated to default to Brave but allow
building either Brave or Chrome/Chromium based on the value of the
`BROWSER` and `BROWSER_VERSION` environment variables.